requirements.txt
absl-py==2.0.0
alembic==1.12.1
aniso8601==9.0.1
apispec==5.2.2
apispec-webframeworks==0.5.2
arrow==1.3.0
astunparse==1.6.3
attrs==23.1.0
Automat==22.10.0
bidict==0.22.1
binaryornot==0.4.4
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.1.1
dask==2023.11.0
deepdiff==6.2.3
distributed==2023.11.0
dnspython==2.4.2
et-xmlfile==1.1.0
Flask==3.0.0
Flask-Cors==4.0.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fsspec==2023.10.0
gast==0.4.0
gevent==23.9.1
gevent-websocket==0.10.1
gitignore-parser==0.1.9
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
hyperlink==21.0.0
idna==3.5
importlib-metadata==6.8.0
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
keras==2.13.1
kthread==0.2.3
libclang==16.0.6
locket==1.0.0
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.28.1
modin==0.25.1
msgpack==1.0.7
networkx==2.8.8
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.9.10
packaging==23.2
pandas==2.1.3
partd==1.4.1
passlib==1.7.4
Pillow==10.1.0
protobuf==4.25.1
psutil==5.9.6
pyarrow==10.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-slugify==8.0.1
python-socketio==5.10.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simple-websocket==0.10.1
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
taipy==3.0.0
taipy-config==3.0.0
taipy-core==3.0.0
taipy-gui==3.0.0
taipy-rest==3.0.0
taipy-templates==3.0.0
tblib==3.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
text-unidecode==1.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
Twisted==23.10.0
twisted-iocpsupport==1.0.4
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
zict==3.0.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
